pyproject.toml
[tool.poetry]
name = "web"
version = "0.1.0"
description = ""
authors = ["Brandon Helms <brandon@golden.co>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
uvicorn = "^0.18.2"
SQLAlchemy = "^1.4.39"
pydantic = "^1.9.1"
poetry-dotenv-plugin = "^0.1.0-alpha.2"
psycopg2 = "^2.9.3"
argon2-cffi = "^21.3.0"
py2neo = "^2021.2.3"
neomodel = "^4.0.8"
PyJWT = "^2.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
